// src/enums/common.ts
export enum WorkTypeEnum {
  GRAPHIC = 1,
  WEB = 0,
}

export enum WebTagEnum {
  ANGULAR = 'Angular',
  HYBRID = 'Hybrid',
  JQUERY = 'JQUERY',
  RWD = 'RWD切版',
  React = 'React',
  UIUX = 'UIUX',
  VUE = 'VUE',
  WORDPRESS = 'WordPress',
}

export enum GraphicEnum {
  ALL = 'ALL',
  BANNER = 'BANNER',
  CIS = 'CIS',
  DM = 'DM',
  GRAPHIC = 'GRAPHIC',
  INFOGRAPHIC = 'INFOGRAPHIC',
  POP = 'POP',
  WEB = 'WEB',
}